+++ src/admin/getpeers.go
@@ -19,7 +19,7 @@ type PeerEntry struct {
 	IPAddress string   `json:"address"`
 	PublicKey string   `json:"key"`
 	Port      uint64   `json:"port"`
-	Priority  uint8    `json:"priority"`
+	Priority  uint64   `json:"priority"`
 	Coords    []uint64 `json:"coords"`
 	Remote    string   `json:"remote"`
 	RXBytes   DataUnit `json:"bytes_recvd"`
@@ -36,7 +36,7 @@ func (a *AdminSocket) getPeersHandler(req *GetPeersRequest, res *GetPeersRespons
 			IPAddress: net.IP(addr[:]).String(),
 			PublicKey: hex.EncodeToString(p.Key),
 			Port:      p.Port,
-			Priority:  p.Priority,
+			Priority:  uint64(p.Priority), // can't be uint8 thanks to gobind
 			Coords:    p.Coords,
 			Remote:    p.Remote,
 			RXBytes:   DataUnit(p.RXBytes),
