+++ src/config/config.go
@@ -44,7 +44,7 @@ type MulticastInterfaceConfig struct {
 	Beacon   bool
 	Listen   bool
 	Port     uint16
-	Priority uint8
+	Priority uint64 // really uint8, but gobind won't export it
 }
 
 // NewSigningKeys replaces the signing keypair in the NodeConfig with a new
